Rename scheduler to executor, move run() to free function
- Rename scheduler.rs → executor.rs; update module doc
- Move RunHandle::run to executor::run() free function; make RunHandle's
  kv/idgen/attachment_config pub(crate) for access from executor
- Environment::run delegates to executor::run
- Update lib.rs, cli, tests, and example to use executor module
